+++ app.php
@@ -6,6 +6,8 @@
 
 use Slim\Factory\AppFactory;
 
+define('KUUSAMO_VERSION', '1.8.2');
+
 function kuusamo()
 {
     // configure services